+++ hypersync-solana-net-types/src/field_selection.rs
@@ -114,6 +114,8 @@ pub enum TokenBalanceField {
     Owner,
     PreAmount,
     PostAmount,
+    PreProgramId,
+    PostProgramId,
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize, Display, EnumString)]